a couple of small fixes to support both i586-mingw32msvc-cc and i686-w64-mingw32-gcc

+++ misc.h
@@ -33,7 +33,7 @@ uint32_t path_from_object_attributes(const OBJECT_ATTRIBUTES *obj,
 int ensure_absolute_path(wchar_t *out, const wchar_t *in, int length);
 
 // imported but for some doesn't show up when #including string.h etc
-int wcsnicmp(const wchar_t *a, const wchar_t *b, int len);
+int wcsnicmp(const wchar_t *a, const wchar_t *b, size_t len);
 int wcsicmp(const wchar_t *a, const wchar_t *b);
 
 int is_shutting_down();
